bump to version 2014.05.10-1

+++ utility2.sh
@@ -56,8 +56,9 @@ shMain () {
 
   ## delete github file
   --db-github-branch-file-delete)
-    SCRIPT="$SCRIPT && $UTILITY2_JS"
-    SCRIPT="$SCRIPT --dbGithubBranchFile=$2"
+    SCRIPT="$SCRIPT && echo deleting"
+    SCRIPT="$SCRIPT https://raw.githubusercontent.com/$GITHUB_OWNER_REPO_BRANCH$2"
+    SCRIPT="$SCRIPT --db-github-branch-file=$2"
     SCRIPT="$SCRIPT --mode-cli=dbGithubBranchFileDelete"
     SCRIPT="$SCRIPT --mode-db-github=$GITHUB_OWNER_REPO_BRANCH"
     SCRIPT="$SCRIPT --mode-silent"
@@ -65,33 +66,38 @@ shMain () {
 
   ## update github file
   --db-github-branch-file-update)
+    SCRIPT="$SCRIPT && echo updating"
+    SCRIPT="$SCRIPT https://raw.githubusercontent.com/$GITHUB_OWNER_REPO_BRANCH$2"
     SCRIPT="$SCRIPT && cat $3 | $UTILITY2_JS"
-    SCRIPT="$SCRIPT --dbGithubBranchFile=$2"
+    SCRIPT="$SCRIPT --db-github-branch-file=$2"
     SCRIPT="$SCRIPT --mode-cli=dbGithubBranchFileUpdate"
     SCRIPT="$SCRIPT --mode-db-github=$GITHUB_OWNER_REPO_BRANCH"
     SCRIPT="$SCRIPT --mode-silent"
+    SCRIPT="$SCRIPT --mode-timeoutDefault=120000"
     ;;
 
   ## install saucelabs
   --saucelabs-install)
-    SAUCELABS_VERSION=$(echo $NODEJS_PROCESS_PLATFORM | perl -ne "s/darwin/osx/i; print")
-    SCRIPT="$SCRIPT && echo installing saucelabs to $UTILITY2_DIR/.install/sc"
-    SCRIPT="$SCRIPT && cd $UTILITY2_DIR"
-    case $NODEJS_PROCESS_PLATFORM in
-    linux)
-      SCRIPT="$SCRIPT && curl -3Ls"
-      SCRIPT="$SCRIPT https://d2nkw87yt5k0to.cloudfront.net/downloads"
-      SCRIPT="$SCRIPT/sc-latest-$SAUCELABS_VERSION.tar.gz"
-      SCRIPT="$SCRIPT | tar -C .install -xzf -"
-      ;;
-    *)
-      SCRIPT="$SCRIPT && curl -3Ls"
-      SCRIPT="$SCRIPT https://d2nkw87yt5k0to.cloudfront.net/downloads"
-      SCRIPT="$SCRIPT/sc-latest-$SAUCELABS_VERSION.zip"
-      SCRIPT="$SCRIPT > .install/sc.zip && unzip -q .install/sc.zip -d .install"
-      ;;
-    esac
-    SCRIPT="$SCRIPT && cp .install/sc-4.1-$SAUCELABS_VERSION/bin/sc .install/sc"
+    if [ ! -f "$UTILITY2_DIR/.install/sc" ]
+      then SAUCELABS_VERSION=$(echo $NODEJS_PROCESS_PLATFORM | perl -ne "s/darwin/osx/i; print")
+      SCRIPT="$SCRIPT && echo installing saucelabs to $UTILITY2_DIR/.install/sc"
+      SCRIPT="$SCRIPT && cd $UTILITY2_DIR"
+      case $NODEJS_PROCESS_PLATFORM in
+      linux)
+        SCRIPT="$SCRIPT && curl -3Ls"
+        SCRIPT="$SCRIPT https://d2nkw87yt5k0to.cloudfront.net/downloads"
+        SCRIPT="$SCRIPT/sc-latest-$SAUCELABS_VERSION.tar.gz"
+        SCRIPT="$SCRIPT | tar -C .install -xzf -"
+        ;;
+      *)
+        SCRIPT="$SCRIPT && curl -3Ls"
+        SCRIPT="$SCRIPT https://d2nkw87yt5k0to.cloudfront.net/downloads"
+        SCRIPT="$SCRIPT/sc-latest-$SAUCELABS_VERSION.zip"
+        SCRIPT="$SCRIPT > .install/sc.zip && unzip -q .install/sc.zip -d .install"
+        ;;
+      esac
+      SCRIPT="$SCRIPT && cp .install/sc-4.1-$SAUCELABS_VERSION/bin/sc .install/sc"
+    fi
     ;;
 
   ## restart saucelabs
@@ -105,16 +111,17 @@ shMain () {
   ## start saucelabs if not already started
   --saucelabs-start)
     SAUCELABS_READY_FILE=/tmp/.saucelabs.ready
-    ## create sauclabs tunnels in a loop until one succeeds
-    SCRIPT="$SCRIPT && while [ ! -f $SAUCELABS_READY_FILE ]"
+    ## start saucelabs only if $SAUCELABS_READY_FILE doesn't exist
+    SCRIPT="$SCRIPT && if [ ! -f $SAUCELABS_READY_FILE ]"
       ## install saucelabs if necessary
-      SCRIPT="$SCRIPT; do if [ ! -f '$UTILITY2_DIR/.install/sc' ]"
-        SCRIPT="$SCRIPT; then $($UTILITY2_SH_ECHO --saucelabs-install)"
-      SCRIPT="$SCRIPT; fi"
-      SCRIPT="$SCRIPT; eval"
-      SCRIPT="$SCRIPT '$UTILITY2_DIR/.install/sc --readyfile $SAUCELABS_READY_FILE &'"
-      SCRIPT="$SCRIPT; sleep 10"
-    SCRIPT="$SCRIPT; done"
+      SCRIPT="$SCRIPT; then $($UTILITY2_SH_ECHO --saucelabs-install)"
+      ## create sauclabs tunnel in a loop until one succeeds
+      SCRIPT="$SCRIPT && while [ ! -f $SAUCELABS_READY_FILE ]"
+        SCRIPT="$SCRIPT; do eval"
+        SCRIPT="$SCRIPT '$UTILITY2_DIR/.install/sc --readyfile $SAUCELABS_READY_FILE &'"
+        SCRIPT="$SCRIPT; sleep 10"
+      SCRIPT="$SCRIPT; done"
+    SCRIPT="$SCRIPT; fi"
     ;;
 
   ## stop saucelabs
@@ -181,11 +188,13 @@ shMain () {
       if [ ! "$CODESHIP" ]
         then exit
       fi
+      HEADLESS_SAUCELABS_URL="https://utility2-unstable.herokuapp.com/test/test.html#modeTest=1"
       ## npm install
       SCRIPT="$SCRIPT && npm install"
+      SCRIPT="$SCRIPT && curl -3Ls $HEADLESS_SAUCELABS_URL > /dev/null"
       ## npm test saucelabs
-      SCRIPT="$SCRIPT && rm -f tmp/test_report.json"
       SCRIPT="$SCRIPT && $($UTILITY2_SH_ECHO --saucelabs-test .install/utility2_saucelabs.all.json)"
+      SCRIPT="$SCRIPT --headless-saucelabs-url=$HEADLESS_SAUCELABS_URL"
       ;;
     *)
       ## npm install
@@ -197,11 +206,12 @@ shMain () {
     ## save EXIT_CODE
     SCRIPT="$SCRIPT; EXIT_CODE=\$?"
     if [ $ARTIFACT_DIR ]
-      ## publish build
-      then SCRIPT="$SCRIPT && ./utility2.sh --utility2-build-publish"
-      SCRIPT="$SCRIPT $ARTIFACT_DIR/latest.$CI_BRANCH"
-      SCRIPT="$SCRIPT && ./utility2.sh --utility2-build-publish"
-      SCRIPT="$SCRIPT $ARTIFACT_DIR/$CI_BUILD_NUMBER.$CI_BRANCH.$CI_COMMIT_ID"
+      ## publish build to $ARTIFACT_DIR/latest.$CI_BRANCH
+      then SCRIPT="$SCRIPT && $($UTILITY2_SH_ECHO --utility2-build-publish $ARTIFACT_DIR/latest.$CI_BRANCH)"
+      ## publish build to $ARTIFACT_DIR/$CI_BUILD_NUMBER.$CI_BRANCH.$CI_COMMIT_ID
+      SCRIPT="$SCRIPT && $($UTILITY2_SH_ECHO --utility2-build-publish $ARTIFACT_DIR/$CI_BUILD_NUMBER.$CI_BRANCH.$CI_COMMIT_ID)"
+      ## exit with error if publish fails
+      SCRIPT="$SCRIPT || exit \$?"
     fi
     SCRIPT="$SCRIPT; exit \$EXIT_CODE"
     ;;
@@ -225,7 +235,8 @@ shMain () {
     SCRIPT="$SCRIPT test_report.badge.svg"
     SCRIPT="$SCRIPT test_report.html"
     SCRIPT="$SCRIPT test_report.json"
-      SCRIPT="$SCRIPT; do if [ -f tmp/\$FILE ]"
+      SCRIPT="$SCRIPT; do [ \$? == 0 ]"
+      SCRIPT="$SCRIPT && if [ -f tmp/\$FILE ]"
         SCRIPT="$SCRIPT; then $($UTILITY2_SH_ECHO --db-github-branch-file-update $2/\$FILE tmp/\$FILE)"
       SCRIPT="$SCRIPT; fi"
     SCRIPT="$SCRIPT; done"
@@ -234,21 +245,22 @@ shMain () {
   ## called by npm run-script build
   --utility2-external-build)
     ## build utility2_external
+    SCRIPT="$SCRIPT && $($UTILITY2_SH_ECHO --utility2-npm-install)"
     SCRIPT="$SCRIPT && ./utility2.js"
     SCRIPT="$SCRIPT --mode-cli=rollupFileList"
     SCRIPT="$SCRIPT --mode-silent"
     SCRIPT="$SCRIPT --rollup-file-list"
     SCRIPT="$SCRIPT=.install/public/utility2_external.browser.css"
     SCRIPT="$SCRIPT,.install/public/utility2_external.browser.js"
-    SCRIPT="$SCRIPT,.install/public/utility2_external.shared.js"
+    SCRIPT="$SCRIPT,.install/public/utility2_external.nodejs.js"
     SCRIPT="$SCRIPT --tmpdir=tmp"
     SCRIPT="$SCRIPT && tar -czvf .install/$UTILITY2_EXTERNAL_TAR_GZ"
     SCRIPT="$SCRIPT .install/public/utility2_external.browser.rollup.css"
     SCRIPT="$SCRIPT .install/public/utility2_external.browser.rollup.js"
     SCRIPT="$SCRIPT .install/public/utility2_external.browser.rollup.min.css"
     SCRIPT="$SCRIPT .install/public/utility2_external.browser.rollup.min.js"
-    SCRIPT="$SCRIPT .install/public/utility2_external.shared.rollup.js"
-    SCRIPT="$SCRIPT .install/public/utility2_external.shared.rollup.min.js"
+    SCRIPT="$SCRIPT .install/public/utility2_external.nodejs.rollup.js"
+    SCRIPT="$SCRIPT .install/public/utility2_external.nodejs.rollup.min.js"
     ;;
 
   ## publish utility2_external on github
@@ -263,7 +275,7 @@ shMain () {
     SCRIPT="$SCRIPT && mkdir -p .install/public"
     SCRIPT="$SCRIPT && ./utility2.js --mode-cli=utility2NpmInstall"
     ## install utility2_external
-    if [ ! -f .install/public/utility2_external.shared.rollup.js ]
+    if [ ! -f .install/public/utility2_external.nodejs.rollup.js ]
       then SCRIPT="$SCRIPT && curl -3Ls"
       SCRIPT="$SCRIPT https://kaizhu256.github.io/blob/utility2_external"
       SCRIPT="$SCRIPT/$UTILITY2_EXTERNAL_TAR_GZ"
